backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
python-docx
python-pptx
pydantic
python-multipart
aiofiles
requests
